feat(order): add Order type definition
Add Order type definition to src/@types/order.d.ts. This type definition
is used to define the shape of an order object.

feat(order): add support for adding new order

Add handleAddOrder function to src/pages/orders/handle-add-order.ts. This
function is used to handle adding new order to the database. It takes in
data from the form, navigate function from react-router-dom, and toast
function from Chakra UI. It then converts the data into an order object
and adds it to the database.

feat(order): add product-list.ts

Add product-list.ts to src/pages/orders. This file contains a list of
products that can be ordered. It is used in AddOrderPage to populate the
product select input.