feat(handle-edit-order.ts): add function to handle editing an order
This commit adds a new file `handle-edit-order.ts` which exports a function `handleEditOrder` that takes in an `EditOrderForm` object, an `id` string, a `NavigateFunction`, a `toast` object, and updates the order with the given `id` in the Firebase Realtime Database with the new data. If there is an error, it displays an error toast with the error message.