feat(utils): add format.ts file to contain utility functions for formatting data
This commit adds a new file named format.ts to the utils folder. This file contains utility functions for formatting data, such as formatAddress, formatPayment, formatDateTime, and formatQtyProduct. These functions are used in various parts of the application to format data before displaying it to the user.

refactor(utils.ts): remove unused functions formatAddress, formatPayment, and formatDateTime to improve code maintainability and readability